backend/requirements.txt
fastapi
uvicorn
python-dotenv
websockets
google-genai
numpy
httpx
deepgram-sdk
elevenlabs
openai
